Refactored module project
SourceFile, VerilogSourceFile and VHDLSourceFile now reside in own
module source_file
